+++ fianchetto/fianchetto/content.py
@@ -1,21 +1,9 @@
-import chess
-from guillotina import configure, content, schema
+from .interfaces import IChessServerManager, IChessGame
+from guillotina import configure, content
 from guillotina.component import get_utility
-from guillotina.directives import index_field
-from guillotina.interfaces import IFolder, IObjectAddedEvent
+from guillotina.interfaces import IObjectAddedEvent
 
-from .interfaces import IChessServerManager
-
-
-class IChessGame(IFolder):
-
-    index_field("white", type="keyword")
-    white = schema.TextLine(title="White player", required=False)
-
-    index_field("black", type="keyword")
-    black = schema.TextLine(title="Black player", required=False)
-
-    moves = schema.List(value_type=schema.TextLine(), default=list())
+import chess
 
 
 @configure.contenttype(type_name="ChessGame", schema=IChessGame)
@@ -28,14 +16,15 @@ class ChessGame(content.Folder):
         :param moves: list e.g. ["f2f4", "e7e6", "g2g4", "d8h4"]
         :return: boolean
         """
-        board = chess.Board()
+        return True
+        # board = chess.Board()
 
-        for move in moves:
-            move_obj = chess.Move.from_uci(move)
-            if move_obj in board.legal_moves:
-                board.push(move_obj)
-            else:
-                return False
+        # for move in moves:
+        #     move_obj = chess.Move.from_uci(move)
+        #     if move_obj in board.legal_moves:
+        #         board.push(move_obj)
+        #     else:
+        #         return False
 
     @property
     def turn(self):
